Move python file reading utils to own file
